+++ server/apiv1/webui.go
@@ -23,15 +23,19 @@ type webUIConfiguration struct {
 		// Allowlist of accepted recipients
 		RecipientAllowlist string
 	}
+
+	// Whether the HTML check has been globally disabled
+	DisableHTMLCheck bool
 }
 
 // WebUIConfig returns configuration settings for the web UI.
-func WebUIConfig(w http.ResponseWriter, r *http.Request) {
+func WebUIConfig(w http.ResponseWriter, _ *http.Request) {
 	// swagger:route GET /api/v1/webui application WebUIConfiguration
 	//
 	// # Get web UI configuration
 	//
 	// Returns configuration settings for the web UI.
+	// Intended for web UI only!
 	//
 	//	Produces:
 	//	- application/json
@@ -49,6 +53,8 @@ func WebUIConfig(w http.ResponseWriter, r *http.Request) {
 		conf.MessageRelay.ReturnPath = config.SMTPRelayConfig.ReturnPath
 		conf.MessageRelay.RecipientAllowlist = config.SMTPRelayConfig.RecipientAllowlist
 	}
+
+	conf.DisableHTMLCheck = config.DisableHTMLCheck
 
 	bytes, _ := json.Marshal(conf)